requirements.txt
python-whois==0.6.7